requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiomysql
pymysql
python-dotenv
jinja2
python-multipart
cryptography
